+++ packages/aws-lambda/src/triggers.js
@@ -216,58 +216,41 @@ function isInvokeFunction(context) {
   return custom && (custom['x-instana-l'] || custom['x-instana-s'] || custom['x-instana-t']);
 }
 
-exports.readTracingHeaders = function readTracingHeaders(event, context) {
-  const tracingHeaders = {};
+exports.readTraceCorrelationData = function readTraceCorrelationData(event, context) {
+  let traceCorrelationData;
 
-  if (event.headers && typeof event.headers === 'object') {
-    tracingHeaders.t = tracingUtil.readAttribCaseInsensitive(event.headers, tracingConstants.traceIdHeaderName);
-    tracingHeaders.s = tracingUtil.readAttribCaseInsensitive(event.headers, tracingConstants.spanIdHeaderName);
-    tracingHeaders.l = tracingUtil.readAttribCaseInsensitive(event.headers, tracingConstants.traceLevelHeaderName);
-    if (hasTracingHeaders(tracingHeaders)) {
-      return tracingHeaders;
-    }
+  traceCorrelationData = readTraceCorrelationFromHttpHeaders(event);
+  if (hasFoundTraceCorrelationData(traceCorrelationData)) {
+    return traceCorrelationData;
   }
 
-  const custom = readClientContextCustom(context);
-  if (custom) {
-    if (custom[tracingConstants.traceLevelHeaderNameLowerCase]) {
-      tracingHeaders.l = custom[tracingConstants.traceLevelHeaderNameLowerCase];
-    }
-    if (custom[tracingConstants.spanIdHeaderNameLowerCase]) {
-      tracingHeaders.s = custom[tracingConstants.spanIdHeaderNameLowerCase];
-    }
-    if (custom[tracingConstants.traceIdHeaderNameLowerCase]) {
-      tracingHeaders.t = custom[tracingConstants.traceIdHeaderNameLowerCase];
-    }
-    if (hasTracingHeaders(tracingHeaders)) {
-      return tracingHeaders;
-    }
+  traceCorrelationData = readTraceCorrelationFromClientContextCustom(context);
+  if (hasFoundTraceCorrelationData(traceCorrelationData)) {
+    return traceCorrelationData;
   }
 
-  const sqsMessageAttributes =
-    Array.isArray(event.Records) && event.Records.length > 0 && event.Records[0].eventSource === 'aws:sqs'
-      ? event.Records[0].messageAttributes
-      : null;
-  if (sqsMessageAttributes) {
-    tracingHeaders.t = readSqsMessageAttributeWithFallback(
-      sqsMessageAttributes,
-      tracingConstants.sqsAttributeNames.TRACE_ID,
-      tracingConstants.sqsAttributeNames.LEGACY_TRACE_ID
-    );
-    tracingHeaders.s = readSqsMessageAttributeWithFallback(
-      sqsMessageAttributes,
-      tracingConstants.sqsAttributeNames.SPAN_ID,
-      tracingConstants.sqsAttributeNames.LEGACY_SPAN_ID
-    );
-    tracingHeaders.l = readSqsMessageAttributeWithFallback(
-      sqsMessageAttributes,
-      tracingConstants.sqsAttributeNames.LEVEL,
-      tracingConstants.sqsAttributeNames.LEGACY_LEVEL
-    );
+  traceCorrelationData = readTraceCorrelationFromSqs(event);
+  if (hasFoundTraceCorrelationData(traceCorrelationData)) {
+    return traceCorrelationData;
   }
 
-  return tracingHeaders;
+  // No trace correlation data has been found, so we return an empty object. This implies that a new trace will be
+  // started.
+  return {};
 };
+
+function readTraceCorrelationFromHttpHeaders(event) {
+  const traceCorrelationData = {};
+  if (event.headers && typeof event.headers === 'object') {
+    traceCorrelationData.t = tracingUtil.readAttribCaseInsensitive(event.headers, tracingConstants.traceIdHeaderName);
+    traceCorrelationData.s = tracingUtil.readAttribCaseInsensitive(event.headers, tracingConstants.spanIdHeaderName);
+    traceCorrelationData.l = tracingUtil.readAttribCaseInsensitive(
+      event.headers,
+      tracingConstants.traceLevelHeaderName
+    );
+  }
+  return traceCorrelationData;
+}
 
 function readClientContextCustom(context) {
   // The Node.js AWS SDK documentation expects for Custom, with capital "C", but the same is not explicitly said for
@@ -276,6 +259,75 @@ function readClientContextCustom(context) {
     (context && context.clientContext && context.clientContext.Custom) ||
     (context && context.clientContext && context.clientContext.custom)
   );
+}
+
+function readTraceCorrelationFromClientContextCustom(context) {
+  const traceCorrelationData = {};
+  const custom = readClientContextCustom(context);
+  if (custom) {
+    if (custom[tracingConstants.traceLevelHeaderNameLowerCase]) {
+      traceCorrelationData.l = custom[tracingConstants.traceLevelHeaderNameLowerCase];
+    }
+    if (custom[tracingConstants.spanIdHeaderNameLowerCase]) {
+      traceCorrelationData.s = custom[tracingConstants.spanIdHeaderNameLowerCase];
+    }
+    if (custom[tracingConstants.traceIdHeaderNameLowerCase]) {
+      traceCorrelationData.t = custom[tracingConstants.traceIdHeaderNameLowerCase];
+    }
+  }
+  return traceCorrelationData;
+}
+
+function readTraceCorrelationFromSqs(event) {
+  let traceCorrelationData = {};
+  if (isSQSTrigger(event)) {
+    const sqsMessageAttributes = event.Records[0].messageAttributes;
+    if (sqsMessageAttributes) {
+      traceCorrelationData = readTraceCorrelationFromSqsAttributes(sqsMessageAttributes);
+      if (hasFoundTraceCorrelationData(traceCorrelationData)) {
+        return traceCorrelationData;
+      }
+    }
+
+    const sqsMessageBody = event.Records[0].body;
+    if (typeof sqsMessageBody === 'string' && sqsMessageBody.startsWith('{')) {
+      try {
+        const parsedSqsMessageBody = JSON.parse(sqsMessageBody);
+        const snsAttributes = parsedSqsMessageBody && parsedSqsMessageBody.MessageAttributes;
+        if (snsAttributes) {
+          traceCorrelationData = readTraceCorrelationFromSqsAttributes(snsAttributes);
+          if (hasFoundTraceCorrelationData(traceCorrelationData)) {
+            return traceCorrelationData;
+          }
+        }
+      } catch (e) {
+        // The attempt to parse the message body as JSON failed, so this is not an SQS message resulting from an SNS
+        // notification (SNS-to-SQS subscription), in which case we are not interested in the body. Ignore the error and
+        // move on.
+      }
+    }
+  }
+  return traceCorrelationData;
+}
+
+function readTraceCorrelationFromSqsAttributes(attributes) {
+  const traceCorrelationData = {};
+  traceCorrelationData.t = readSqsMessageAttributeWithFallback(
+    attributes,
+    tracingConstants.sqsAttributeNames.TRACE_ID,
+    tracingConstants.sqsAttributeNames.LEGACY_TRACE_ID
+  );
+  traceCorrelationData.s = readSqsMessageAttributeWithFallback(
+    attributes,
+    tracingConstants.sqsAttributeNames.SPAN_ID,
+    tracingConstants.sqsAttributeNames.LEGACY_SPAN_ID
+  );
+  traceCorrelationData.l = readSqsMessageAttributeWithFallback(
+    attributes,
+    tracingConstants.sqsAttributeNames.LEVEL,
+    tracingConstants.sqsAttributeNames.LEGACY_LEVEL
+  );
+  return traceCorrelationData;
 }
 
 function readSqsMessageAttributeWithFallback(messageAttributes, key, keyFallback) {
@@ -287,12 +339,13 @@ function readSqsMessageAttributeWithFallback(messageAttributes, key, keyFallback
 
 function readSqsStringMessageAttribute(messageAttributes, key) {
   const attribute = tracingUtil.readAttribCaseInsensitive(messageAttributes, key);
-  if (attribute && attribute.stringValue) {
-    return attribute.stringValue;
+  // attribute.stringValue is used by SQS message attributes, attribute.Value is used by SNS-to-SQS.
+  if (attribute && (attribute.stringValue || attribute.Value)) {
+    return attribute.stringValue || attribute.Value;
   }
   return null;
 }
 
-function hasTracingHeaders(tracingHeaders) {
-  return tracingHeaders.t || tracingHeaders.s || tracingHeaders.l;
+function hasFoundTraceCorrelationData(traceCorrelationData) {
+  return traceCorrelationData.t || traceCorrelationData.s || traceCorrelationData.l;
 }