+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-- [AWS Lambda] Add trace correlation support for Lambda invokations triggered by SQS messages.
+- [AWS Lambda] Add trace correlation support for Lambda invocations triggered by SQS messages (including SNS-to-SQS).
 
 ## 1.127.0
 - AWS SDK v3: Instrumentation of DynamoDB